refactor(c_array): implement c_array function in .c file to prevent the multiple declartion
