+++ src/c_array.h
@@ -189,33 +189,13 @@ typedef double var_t;
         double*: cmpfunc_double                         \
     )
 
-int cmpfunc_int(const void * a, const void * b) {
-   return ( *(int*)a - *(int*)b );
-}
+int cmpfunc_int(const void * a, const void * b);
 
-int cmpfunc_long(const void * a, const void * b) {
-   return ( *(long long*)a - *(long long*)b );
-}
+int cmpfunc_long(const void * a, const void * b);
 
-int cmpfunc_float(const void * a, const void * b) {
-    float ret = *(float*)a - *(float*)b;
-    if(ret > 0){
-        return 1;
-    } else if (ret < 0) {
-        return -1;
-    }
-    return 0;
-}
+int cmpfunc_float(const void * a, const void * b);
 
-int cmpfunc_double(const void * a, const void * b) {
-    double ret = *(double*)a - *(double*)b;
-    if(ret > 0){
-        return 1;
-    } else if (ret < 0) {
-        return -1;
-    }
-    return 0;
-}
+int cmpfunc_double(const void * a, const void * b);
 
 // -----------------------------------------------------------------------
 /*                           Array msort                                */
@@ -279,21 +259,13 @@ int cmpfunc_double(const void * a, const void * b) {
     }                                       \
     return sum;                             \
 
-int c_array_sum_int(int* arr, int size) {
-    c_array_sum_process(arr, size);
-}
+int c_array_sum_int(int* arr, int size);
 
-long long c_array_sum_long(long long* arr, int size) {
-    c_array_sum_process(arr, size);
-}
+long long c_array_sum_long(long long* arr, int size);
 
-float c_array_sum_float(float* arr, int size) {
-    c_array_sum_process(arr, size);
-}
+float c_array_sum_float(float* arr, int size);
 
-double c_array_sum_double(double* arr, int size) {
-    c_array_sum_process(arr, size);
-}
+double c_array_sum_double(double* arr, int size);
 
 // -----------------------------------------------------------------------
 /*                          Array mean value                            */
@@ -310,21 +282,13 @@ double c_array_sum_double(double* arr, int size) {
     mean_t mean = (mean_t)(sum) / (size);       \
     return mean;
 
-mean_t c_array_mean_int(int* arr, int size, int sum) {
-    c_array_mean_process(arr, size, sum);
-}
+mean_t c_array_mean_int(int* arr, int size, int sum);
 
-mean_t c_array_mean_long_long(long long* arr, int size, long long sum) {
-    c_array_mean_process(arr, size, sum);
-}
+mean_t c_array_mean_long_long(long long* arr, int size, long long sum);
 
-mean_t c_array_mean_float(float* arr, int size, float sum) {
-    c_array_mean_process(arr, size, sum);
-}
+mean_t c_array_mean_float(float* arr, int size, float sum);
 
-mean_t c_array_mean_double(double* arr, int size, double sum) {
-    c_array_mean_process(arr, size, sum);
-}
+mean_t c_array_mean_double(double* arr, int size, double sum);
 
 // -----------------------------------------------------------------------
 /*                     Array standard deviation                         */
@@ -351,45 +315,21 @@ mean_t c_array_mean_double(double* arr, int size, double sum) {
         var += pow((arr)[i] - (mean), 2);      \
     }                                          \
 
-var_t c_array_var_int(int* arr, int size, mean_t mean) {
-    c_array_var_process(arr, size, mean);
-    return (var / (size));
-}
+var_t c_array_var_int(int* arr, int size, mean_t mean);
 
-var_t c_array_var_long_long(long long* arr, int size, mean_t mean) {
-    c_array_var_process(arr, size, mean);
-    return (var / (size));
-}
+var_t c_array_var_long_long(long long* arr, int size, mean_t mean);
 
-var_t c_array_var_float(float* arr, int size, mean_t mean) {
-    c_array_var_process(arr, size, mean);
-    return (var / (size));
-}
+var_t c_array_var_float(float* arr, int size, mean_t mean);
 
-var_t c_array_var_double(double* arr, int size, mean_t mean) {
-    c_array_var_process(arr, size, mean);
-    return (var / (size));
-}
+var_t c_array_var_double(double* arr, int size, mean_t mean);
 
-std_t c_array_std_int(int* arr, int size, mean_t mean) {
-    c_array_var_process(arr, size, mean);
-    return sqrt(var / (size));
-}
+std_t c_array_std_int(int* arr, int size, mean_t mean);
 
-std_t c_array_std_long_long(long long* arr, int size, mean_t mean) {
-    c_array_var_process(arr, size, mean);
-    return sqrt(var / (size));
-}
+std_t c_array_std_long_long(long long* arr, int size, mean_t mean);
 
-std_t c_array_std_float(float* arr, int size, mean_t mean) {
-    c_array_var_process(arr, size, mean);
-    return sqrt(var / (size));
-}
+std_t c_array_std_float(float* arr, int size, mean_t mean);
 
-std_t c_array_std_double(double* arr, int size, mean_t mean) {
-    c_array_var_process(arr, size, mean);
-    return sqrt(var / (size));
-}
+std_t c_array_std_double(double* arr, int size, mean_t mean);
 
 // -----------------------------------------------------------------------
 /*                          Array Min and Max                           */
@@ -448,37 +388,21 @@ std_t c_array_std_double(double* arr, int size, mean_t mean) {
         return min;                                                         \
     }                                                                       \
 
-int c_array_max_int(int* arr, int size) {
-    c_array_maxmin_process(arr, size, "max");
-}
+int c_array_max_int(int* arr, int size);
 
-long long c_array_max_long_long(long long* arr, int size) {
-    c_array_maxmin_process(arr, size, "max");
-}
+long long c_array_max_long_long(long long* arr, int size);
 
-float c_array_max_float(float* arr, int size) {
-    c_array_maxmin_process(arr, size, "max");
-}
+float c_array_max_float(float* arr, int size);
 
-double c_array_max_double(double* arr, int size) {
-    c_array_maxmin_process(arr, size, "max");
-}
+double c_array_max_double(double* arr, int size);
 
-int c_array_min_int(int* arr, int size) {
-    c_array_maxmin_process(arr, size, "min");
-}
+int c_array_min_int(int* arr, int size);
 
-long long c_array_min_long_long(long long* arr, int size) {
-    c_array_maxmin_process(arr, size, "min");
-}
+long long c_array_min_long_long(long long* arr, int size);
 
-float c_array_min_float(float* arr, int size) {
-    c_array_maxmin_process(arr, size, "min");
-}
+float c_array_min_float(float* arr, int size);
 
-double c_array_min_double(double* arr, int size) {
-    c_array_maxmin_process(arr, size, "min");
-}
+double c_array_min_double(double* arr, int size);
 
 // -----------------------------------------------------------------------
 /*                            Arrary utils                              */
